requirements.txt
numpy
pandas
python-dotenv
pymysql
mysql-connector-python
scikit-learn
# -e .